consolidating utilities into `:compiler-k2-utils`
