+++ compiler-k2-utils/src/main/kotlin/com/squareup/anvil/compiler/k2/utils/names/ClassIds.kt
@@ -93,6 +93,8 @@ private fun classId(
   isLocal = isLocal,
 )
 
+public fun FqName.classId(): ClassId = ClassId.topLevel(this)
+
 /**
  * ```
  *  given: `com.example.SomeClass
