+++ compiler-k2-utils/src/main/kotlin/com/squareup/anvil/compiler/k2/utils/names/Names.kt
@@ -3,7 +3,7 @@ package com.squareup.anvil.compiler.k2.utils.names
 import org.jetbrains.kotlin.name.Name
 
 public object Names {
-  public val scope: Name = Name.identifier("scope")
-  public val modules: Name = Name.identifier("modules")
   public val dependencies: Name = Name.identifier("dependencies")
+  public val modules: Name = Name.identifier("modules")
+  public val scope: Name = Name.identifier("scope")
 }
