+++ compiler-k2-utils/src/main/kotlin/com/squareup/anvil/compiler/k2/utils/stdlib/CollectionUtils.kt
@@ -1,4 +1,4 @@
-package com.squareup.anvil.compiler.k2.utils
+package com.squareup.anvil.compiler.k2.utils.stdlib
 
 private val truePredicate: (Any?) -> Boolean = { true }
