+++ compiler-k2/api/compiler-k2.api
@@ -70,7 +70,3 @@ public final class com/squareup/anvil/compiler/k2/ir/GeneratedDeclarationsIrBody
 	public fun resolveSymbol (Lorg/jetbrains/kotlin/ir/symbols/IrSymbol;Lorg/jetbrains/kotlin/ir/builders/TranslationPluginContext;)Lorg/jetbrains/kotlin/ir/declarations/IrDeclaration;
 }
 
-public final class com/squareup/anvil/compiler/k2/util/UtilsKt {
-	public static final fun classId (Lorg/jetbrains/kotlin/name/FqName;)Lorg/jetbrains/kotlin/name/ClassId;
-}
-
